Added check infrastructure to the test runner.
For every unit test module that will be loaded, a file named 'checks.py' is
looked up in its package.  If such a file exists, it is imported and its
test_hooks() function gets called.  This function should return a list of
hooks -- objects that have the following methods:

  def startTest(self, test):
      """Hook that gets called before running a test."""
  
  def stopTest(self, test):
      """Hook that gets called after running a test."""

Added a sample checks module for the schooltool package.  It checks that
unit tests do not change the global registries in schooltool.component, and
that unit tests do not leave changes in the transaction.