Improved encapsulation of the event logging facet.
Changed PersistentList to OOBTree, which makes things faster, and makes
the database grow less. The results are comparable to using a PersistentList
that is constrained to being a maximum of 25 items in length.

In general, though, keeping the event logger on all the time bloats the
database by 300% and has a 30% speed penalty while importing data.

It may be better to keep the logger turned off by default.